Move ConnectionRegistry and ClientState to ClientConnectionManager.ts Remove unused imports and properties Rename ClientState to ConnectionState
