+++ src/invocation/InvocationService.ts
@@ -45,7 +45,7 @@ import {ClientConfig} from '../config';
 import {ListenerService} from '../listener/ListenerService';
 import {ClientErrorFactory} from '../protocol/ErrorFactory';
 import {LifecycleService} from '../LifecycleService';
-import {ConnectionRegistry} from '../network/ConnectionRegistry';
+import {ConnectionRegistry} from '../network/ClientConnectionManager';
 
 const MAX_FAST_INVOCATION_COUNT = 5;
 const PROPERTY_INVOCATION_RETRY_PAUSE_MILLIS = 'hazelcast.client.invocation.retry.pause.millis';
