+++ src/network/ClientConnectionManager.ts
@@ -32,7 +32,7 @@ import {
     LoadBalancer,
     AddressImpl,
     Addresses,
-    MemberImpl
+    MemberImpl, ClientOfflineError, IOError
 } from '../core';
 import {lookupPublicAddress} from '../core/MemberInfo';
 import {ClientConnection} from './ClientConnection';
@@ -52,7 +52,7 @@ import {ILogger} from '../logging/ILogger';
 import {HeartbeatManager} from './HeartbeatManager';
 import {UuidUtil} from '../util/UuidUtil';
 import {WaitStrategy} from './WaitStrategy';
-import {ReconnectMode} from '../config/ConnectionStrategyConfig';
+import {ConnectionStrategyConfig, ReconnectMode} from '../config/ConnectionStrategyConfig';
 import {ClientConfig, ClientConfigImpl} from '../config/Config';
 import {LifecycleState, LifecycleServiceImpl, LifecycleService} from '../LifecycleService';
 import {ClientMessage} from '../protocol/ClientMessage';
@@ -69,7 +69,6 @@ import {AddressProvider} from '../connection/AddressProvider';
 import {ClientFailoverConfig} from '../config';
 import {ClusterService} from '../invocation/ClusterService';
 import {SerializationService} from '../serialization/SerializationService';
-import {ConnectionRegistry, ClientState} from './ConnectionRegistry';
 
 const CONNECTION_REMOVED_EVENT_NAME = 'connectionRemoved';
 const CONNECTION_ADDED_EVENT_NAME = 'connectionAdded';
@@ -79,6 +78,215 @@ export const CLIENT_TYPE = 'NJS';
 const SERIALIZATION_VERSION = 1;
 const SET_TIMEOUT_MAX_DELAY = 2147483647;
 const BINARY_PROTOCOL_VERSION = Buffer.from('CP2');
+
+export enum ConnectionState {
+    /**
+     * Clients start with this state. Once a client connects to a cluster,
+     * it directly switches to {@link INITIALIZED_ON_CLUSTER} instead of
+     * {@link CONNECTED_TO_CLUSTER} because on startup a client has no
+     * local state to send to the cluster.
+     */
+    INITIAL = 0,
+
+    /**
+     * When a client switches to a new cluster, it moves to this state.
+     * It means that the client has connected to a new cluster but not sent
+     * its local state to the new cluster yet.
+     */
+    CONNECTED_TO_CLUSTER = 1,
+
+    /**
+     * When a client sends its local state to the cluster it has connected,
+     * it switches to this state. When a client loses all connections to
+     * the current cluster and connects to a new cluster, its state goes
+     * back to {@link CONNECTED_TO_CLUSTER}.
+     * <p>
+     * Invocations are allowed in this state.
+     */
+    INITIALIZED_ON_CLUSTER = 2,
+}
+
+export interface ConnectionRegistry {
+    /**
+     * Returns if the registry active
+     * @return Return true if the registry is active, false otherwise
+     */
+    isActive(): boolean;
+
+    /**
+     * Returns if the registry empty
+     * @return true if the registry empty, false otherwise
+     */
+    isEmpty(): boolean;
+
+    /**
+     * Returns connection by UUID
+     * @param uuid UUID that identifies the connection
+     * @return ClientConnection if there is a connection with the UUID, undefined otherwise
+     */
+    getConnection(uuid: UUID): ClientConnection | undefined;
+
+    /**
+     * Returns all active connections in the registry
+     * @return Array of ClientConnection objects
+     */
+    getConnections(): ClientConnection[];
+
+    /**
+     * Returns a random connection from active connections
+     * @return ClientConnection if there is at least one connection, otherwise null
+     */
+    getRandomConnection(): ClientConnection | null;
+
+    /**
+     * Runs the given function for every active connection
+     * @param fn A function that takes ClientConnection as arg, does not return anything
+     */
+    forEachConnection(fn: (conn: ClientConnection) => void): void;
+
+    /**
+     * Returns if invocation allowed. Invocation is allowed only if client state is {@link INITIALIZED_ON_CLUSTER}
+     * and there is at least one active connection.
+     * @return Error if invocation is not allowed, null otherwise
+     */
+    checkIfInvocationAllowed(): Error | null;
+
+    /**
+     * Delete connection from active connections
+     * @param uuid UUID that identifies the connection
+     */
+    deleteConnection(uuid: UUID): void;
+
+    /**
+     * Returns client state.
+     * @return ConnectionState enum value
+     */
+    getClientState(): ConnectionState;
+
+    /**
+     * Sets the client state
+     * @param clientState
+     */
+    setClientState(clientState: ConnectionState): void;
+
+    /**
+     * Adds or updates a client connection by uuid
+     * @param uuid UUID to identify the connection
+     * @param connection the ClientConnection to set
+     */
+    setConnection(uuid: UUID, connection: ClientConnection): void
+
+    /**
+     * Sets the active property of the ConnectionRegistry to true
+     */
+    activate(): void;
+
+    /**
+     * Sets the active property of the ConnectionRegistry to false
+     */
+    deactivate(): void;
+}
+
+export class ConnectionRegistryImpl implements ConnectionRegistry {
+
+    private active = false;
+    private readonly activeConnections = new Map<string, ClientConnection>();
+    private readonly loadBalancer: LoadBalancer;
+    private clientState = ConnectionState.INITIAL;
+    private readonly smartRoutingEnabled: boolean;
+    private readonly asyncStart: boolean;
+    private readonly reconnectMode: ReconnectMode;
+
+    constructor(connectionStrategy: ConnectionStrategyConfig) {
+        this.asyncStart = connectionStrategy.asyncStart;
+        this.reconnectMode = connectionStrategy.reconnectMode;
+    }
+
+    isActive(): boolean {
+        return this.active;
+    }
+
+    getClientState(): ConnectionState {
+        return this.clientState;
+    }
+
+    isEmpty(): boolean {
+        return this.activeConnections.size === 0;
+    }
+
+    getConnections(): ClientConnection[] {
+        return Array.from(this.activeConnections.values());
+    }
+
+    getConnection(uuid: UUID): ClientConnection | undefined {
+        return this.activeConnections.get(uuid.toString());
+    }
+
+    getRandomConnection(): ClientConnection {
+        if (this.smartRoutingEnabled) {
+            const member = this.loadBalancer.next();
+            if (member != null) {
+                const connection = this.getConnection(member.uuid);
+                if (connection != null) {
+                    return connection;
+                }
+            }
+        }
+
+        const iterator = this.activeConnections.values();
+        const next = iterator.next();
+        if (!next.done) {
+            return next.value;
+        } else {
+            return null;
+        }
+    }
+
+    forEachConnection(fn: (conn: ClientConnection) => void): void {
+        this.activeConnections.forEach(fn);
+    }
+
+    checkIfInvocationAllowed(): Error {
+        const state = this.clientState;
+        if (state === ConnectionState.INITIALIZED_ON_CLUSTER && this.activeConnections.size > 0) {
+            return null;
+        }
+
+        let error: Error;
+        if (state === ConnectionState.INITIAL) {
+            if (this.asyncStart) {
+                error = new ClientOfflineError();
+            } else {
+                error = new IOError('No connection found to cluster since the client is starting.');
+            }
+        } else if (this.reconnectMode === ReconnectMode.ASYNC) {
+            error = new ClientOfflineError();
+        } else {
+            error = new IOError('No connection found to cluster.');
+        }
+        return error;
+    }
+
+    deleteConnection(uuid: UUID): void {
+        this.activeConnections.delete(uuid.toString());
+    }
+
+    setConnection(uuid: UUID, connection: ClientConnection): void {
+        this.activeConnections.set(uuid.toString(), connection);
+    }
+
+    setClientState(clientState: ConnectionState): void {
+        this.clientState = clientState;
+    }
+
+    activate(): void {
+        this.active = true;
+    }
+
+    deactivate(): void {
+        this.active = false;
+    }
+}
 
 interface ClientForClientConnectionManager {
     onClusterChange(): void;
@@ -97,8 +305,6 @@ interface ClientForClientConnectionManager {
 export class ClientConnectionManager extends EventEmitter {
 
     private connectionIdCounter = 0;
-    private alive = false;
-
     private readonly logger: ILogger;
     private readonly client: ClientForClientConnectionManager;
     private readonly labels: string[];
@@ -175,10 +381,10 @@ export class ClientConnectionManager extends EventEmitter {
     }
 
     start(): Promise<void> {
-        if (this.alive) {
+        if (this.connectionRegistry.isActive()) {
             return Promise.resolve();
         }
-        this.alive = true;
+        this.connectionRegistry.activate();
 
         this.heartbeatManager.start(this.invocationService);
         return this.connectToCluster();
@@ -197,11 +403,11 @@ export class ClientConnectionManager extends EventEmitter {
     }
 
     shutdown(): void {
-        if (!this.alive) {
+        if (!this.connectionRegistry.isActive()) {
             return;
         }
 
-        this.alive = false;
+        this.connectionRegistry.deactivate();
         if (this.reconnectToMembersTask !== undefined) {
             cancelRepetitionTask(this.reconnectToMembersTask);
         }
@@ -226,7 +432,7 @@ export class ClientConnectionManager extends EventEmitter {
     }
 
     isAlive(): boolean {
-        return this.alive;
+        return this.connectionRegistry.isActive();
     }
 
     getClientUuid(): UUID {
@@ -332,7 +538,7 @@ export class ClientConnectionManager extends EventEmitter {
             this.logger.info('ConnectionManager', 'Removed connection to endpoint: '
                 + endpoint + ':' + memberUuid + ', connection: ' + connection);
             if (this.connectionRegistry.isEmpty()) {
-                if (this.connectionRegistry.getClientState() === ClientState.INITIALIZED_ON_CLUSTER) {
+                if (this.connectionRegistry.getClientState() === ConnectionState.INITIALIZED_ON_CLUSTER) {
                     this.emitLifecycleEvent(LifecycleState.DISCONNECTED);
                 }
                 this.triggerClusterReconnection();
@@ -808,10 +1014,10 @@ export class ClientConnectionManager extends EventEmitter {
         if (this.connectionRegistry.isEmpty()) {
             this.clusterId = newClusterId;
             if (clusterIdChanged) {
-                this.connectionRegistry.setClientState(ClientState.CONNECTED_TO_CLUSTER);
+                this.connectionRegistry.setClientState(ConnectionState.CONNECTED_TO_CLUSTER);
                 this.initializeClientOnCluster(newClusterId);
             } else {
-                this.connectionRegistry.setClientState(ClientState.INITIALIZED_ON_CLUSTER);
+                this.connectionRegistry.setClientState(ConnectionState.INITIALIZED_ON_CLUSTER);
                 this.emitLifecycleEvent(LifecycleState.CONNECTED);
             }
         }
@@ -891,7 +1097,7 @@ export class ClientConnectionManager extends EventEmitter {
                     this.logger.trace('ConnectionManager', 'Client state is sent to cluster: '
                         + targetClusterId);
 
-                    this.connectionRegistry.setClientState(ClientState.INITIALIZED_ON_CLUSTER);
+                    this.connectionRegistry.setClientState(ConnectionState.INITIALIZED_ON_CLUSTER);
                     this.emitLifecycleEvent(LifecycleState.CONNECTED);
                 } else {
                     this.logger.warn('ConnectionManager', 'Cannot set client state to initialized on '
