+++ src/proxy/PartitionSpecificProxy.ts
@@ -23,10 +23,9 @@ import {ProxyManager} from './ProxyManager';
 import {PartitionService} from '../PartitionService';
 import {InvocationService} from '../invocation/InvocationService';
 import {SerializationService} from '../serialization/SerializationService';
-import {ClientConnectionManager} from '../network/ClientConnectionManager';
+import {ConnectionRegistry} from '../network/ClientConnectionManager';
 import {ListenerService} from '../listener/ListenerService';
 import {ClusterService} from '../invocation/ClusterService';
-import {ConnectionRegistry} from '../network/ConnectionRegistry';
 
 /** @internal */
 export class PartitionSpecificProxy extends BaseProxy {
@@ -42,7 +41,6 @@ export class PartitionSpecificProxy extends BaseProxy {
         partitionService: PartitionService,
         invocationService: InvocationService,
         serializationService: SerializationService,
-        connectionManager: ClientConnectionManager,
         listenerService: ListenerService,
         clusterService: ClusterService,
         connectionRegistry: ConnectionRegistry
@@ -56,7 +54,6 @@ export class PartitionSpecificProxy extends BaseProxy {
             partitionService,
             invocationService,
             serializationService,
-            connectionManager,
             listenerService,
             clusterService,
             connectionRegistry
